# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ticket Analyzer AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with AI about tickets (with conversation memory)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /ai/analyze/{uploadId}:
    post:
      operationId: analyzeUpload
      tags: [ai]
      summary: Auto-analyze a CSV upload and generate AI insights
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
  /ai/report/{uploadId}:
    get:
      operationId: getReport
      tags: [ai]
      summary: Generate an exportable Markdown report
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Markdown report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
  /ai/anomalies:
    get:
      operationId: detectAnomalies
      tags: [ai]
      summary: Detect anomalies and unusual patterns in tickets
      parameters:
        - name: upload_id
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Anomaly detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnomalyResult"
  /ai/categorize/{uploadId}:
    post:
      operationId: categorizeTickets
      tags: [ai]
      summary: Auto-categorize uncategorized tickets using AI
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Categorization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategorizationResult"
...
components:
  schemas:
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string
        timestamp:
          type: ["string", "null"]
    ChatInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
        upload_id:
          type: ["integer", "null"]
        conversation_id:
          type: ["string", "null"]
    ChatResponse:
      type: object
      required: [answer, conversation_id]
      properties:
        answer:
          type: string
        conversation_id:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    AnalysisResult:
      type: object
      required: [summary, key_findings, recommendations]
      properties:
        summary:
          type: string
        key_findings:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        top_categories:
          type: array
          items:
            type: string
        risk_level:
          type: string
        anomalies_detected:
          type: integer
    ReportResult:
      type: object
      required: [markdown, generated_at]
      properties:
        markdown:
          type: string
        generated_at:
          type: string
    AnomalyItem:
      type: object
      required: [type, description, severity]
      properties:
        type:
          type: string
        description:
          type: string
        severity:
          type: string
        affected_tickets:
          type: ["integer", "null"]
        affected_category:
          type: ["string", "null"]
    AnomalyResult:
      type: object
      required: [anomalies, summary]
      properties:
        anomalies:
          type: array
          items:
            $ref: "#/components/schemas/AnomalyItem"
        summary:
          type: string
        risk_score:
          type: integer
    CategorySuggestion:
      type: object
      required: [ticket_id, suggested_category, confidence]
      properties:
        ticket_id:
          type: integer
        suggested_category:
          type: string
        confidence:
          type: number
    CategorizationResult:
      type: object
      required: [suggestions, applied]
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/CategorySuggestion"
        applied:
          type: integer
...